# routes/sexo/rotas_sexo.js
const express = require('express');
const router = express.Router();

const controllerSexo = require('../../controller/sexo/controller_sexo.js');

router.post('/', async function (request, response) {
    let dados = request.body;
    let contentType = request.headers['content-type'];

    let result = await controllerSexo.inserirNovoSexo(dados, contentType);

    response.status(result.status_code);
    response.json(result);
});

router.get('/', async function (request, response) {
    let result = await controllerSexo.listarTodosSexos();

    response.status(result.status_code);
    response.json(result);
});

router.get('/:id', async function (request, response) {
    let id = request.params.id;

    let result = await controllerSexo.buscarSexo(id);

    response.status(result.status_code);
    response.json(result);
});

router.put('/:id', async function (request, response) {
    let dados = request.body;
    let id = request.params.id;
    let contentType = request.headers['content-type'];

    let result = await controllerSexo.atualizarSexo(dados, id, contentType);

    response.status(result.status_code);
    response.json(result);
});

router.delete('/:id', async function (request, response) {
    let id = request.params.id;

    let result = await controllerSexo.deletarSexo(id);

    response.status(result.status_code);
    response.json(result);
});

module.exports = router;