# swagger.yaml
openapi: 3.0.3
info:
  title: Super API gerenciadora de filmes
  description: API para gerenciamento de filmes, contendo controle completo de elenco, equipe técnica, pessoas, premiações e mídias associadas.
  contact:
    email: gabriel4321uo@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1/senai/locadora
paths:
...
  /tipo_atuacao/:
    get:
      tags:
        - tipo_atuacao
      summary: Rota para obter todos os tipos atuação.
      description: Rota para obter todos os tipos atuação.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      tipo_atuacao:
                        type: array
                        items:
                          $ref: '#/components/schemas/tipo_atuacao'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    post:
      tags:
        - tipo_atuacao
      summary: Rota usada para inserir um tipo de atuação.
      description: Rota usada para inserir um tipo de atuação.
      responses:
        '201':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  response:
                    type: object
                    properties:
                      tipo_atuacao:
                        type: array
                        items:
                          $ref: '#/components/schemas/tipo_atuacao'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tipo_atuacao'
  /tipo_atuacao/:id:
    get:
      tags:
        - tipo_atuacao
      summary: Rota para obter um tipo de atuação pelo id.
      description: Rota para obter um tipo de atuação pelo id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do tipo de atuação.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      tipo_atuacao:
                        type: array
                        items:
                          $ref: '#/components/schemas/tipo_atuacao'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    put:
      tags:
        - tipo_atuacao
      summary: Rota usada para atualizar um tipo de atuação
      description: Rota usada para atualizar um tipo de atuação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do tipo de atuação.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      tipo_atuacao:
                        type: array
                        items:
                          $ref: '#/components/schemas/tipo_atuacao'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tipo_atuacao'
    delete:
      tags:
        - tipo_atuacao
      summary: Rota usada para deletar um tipo de atuação.
      description: Rota usada para deletar um tipo de atuação.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do tipo de atuação.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUCCESS_DELETED_ITEM'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
...
  /sexo/:
    get:
      tags:
        - sexo
      summary: Rota para obter todos os sexos.
      description: Rota para obter todos os sexos.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      sexo:
                        type: array
                        items:
                          $ref: '#/components/schemas/Sexo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    post:
      tags:
        - sexo
      summary: Rota para inserir um sexo.
      description: Rota para inserir um sexo.
      responses:
        '201':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  response:
                    type: object
                    properties:
                      sexo:
                        type: array
                        items:
                          $ref: '#/components/schemas/Sexo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
  /sexo/:id:
    get:
      tags:
        - sexo
      summary: Rota para obter um sexo pelo id.
      description: Rota para obter um sexo pelo id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do sexo a obter.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      sexo:
                        type: array
                        items:
                          $ref: '#/components/schemas/Sexo'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    put:
      tags:
        - sexo
      summary: Rota para atualizar um sexo.
      description: Rota para atualizar um sexo.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do sexo a atualizar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      sexo:
                        type: array
                        items:
                          $ref: '#/components/schemas/Sexo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
    delete:
      tags:
        - sexo
      summary: Rota para deletar um sexo.
      description: Rota para deletar um sexo.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do sexo a deletar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUCCESS_DELETED_ITEM'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
...
components:
  schemas:
    ERROR_BAD_REQUEST:
      type: object
      properties:
        status:
          type: boolean
          description: status false
        status_code:
          type: integer
          format: int64
          description: '400'
        message:
          type: string
          description: Não foi possivel processar a requisição devido a erros de entrada de dados 😣.
        image:
          type: string
          description: https://http.dog/400.jpg
    SUCCESS_DELETED_ITEM:
      type: object
      properties:
        status:
          type: boolean
          description: 'true'
        status_code:
          type: integer
          format: int64
          description: '200'
        message:
          type: string
          description: Item excluido com sucesso 🥳
        image:
          type: string
          description: https://http.dog/200.jpg
    ERROR_INTERNAL_SERVER_MODEL:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '500'
        message:
          type: string
          description: Não foi possível processar a requisição devido a um erro interno no servidor. 🤖 [MODEL].
        image:
          type: string
          description: https://http.dog/500.jpg
    ERROR_INTERNAL_SERVER_CONTROLLER:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '500'
        message:
          type: string
          description: Não foi possível processar a requisição devido a um erro interno no servidor. 🤖 [CONTROLLER].
        image:
          type: string
          description: https://http.dog/500.jpg
    ERROR_CONTENT_TYPE:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '415'
        message:
          type: string
          description: Não foi possível processar a requisição, pois o formato de dados ecaminhado não é suportado pelo servidor, apenas deve ser utilizado JSON. 🤖
        image:
          type: string
          description: https://http.dog/415.jpg
    ERROR_NOT_FOUND:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '404'
        message:
          type: string
          description: Não foram encontrados dados para retorno. 🗺️❌
        image:
          type: string
          description: https://http.dog/404.jpg
...
    tipo_atuacao:
      required:
        - tipo
        - descricao
      type: object
      properties:
        tipo:
          type: string
          description: Nome do tipo de atuação feita.
        descricao:
          type: string
          description: Descrição do tipo de atuação.
...
    Sexo:
      required:
        - sexo
        - sigla
      type: object
      properties:
        sexo:
          type: string
          description: Nome do sexo a inserir.
        sigla:
          type: string
          description: Sigla do sexo a inserir.
...